% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{from_json}
\alias{from_json}
\title{parse a JSON string}
\usage{
from_json(json, strip_comments = FALSE, ...)
}
\arguments{
\item{json}{a JSON string.}

\item{strip_comments}{whether to strip comments, cf \code{\link[=strip_comments]{strip_comments()}}}

\item{...}{passed to jsonlite::fromJSON}
}
\value{
the decoded data structure, or NA if json is NA, or NULL if json is NULL
}
\description{
parse a JSON string
}
