% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{read_json}
\alias{read_json}
\title{read a JSON file}
\usage{
read_json(path, ...)
}
\arguments{
\item{path}{the path of the JSON config file}

\item{...}{
  Arguments passed on to \code{\link[=from_json]{from_json}}
  \describe{
    \item{\code{json}}{a JSON string.}
    \item{\code{strip_comments}}{whether to strip comments, cf \code{\link[=strip_comments]{strip_comments()}}}
  }}
}
\value{
the parsed JSON as a R data structure
}
\description{
read a JSON file
}
