% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{to_json}
\alias{to_json}
\title{currently just a wrapper around jsonlite::toJSON}
\usage{
to_json(..., pretty = TRUE)
}
\arguments{
\item{...}{passed to jsonlite::fromJSON}

\item{pretty}{adds indentation whitespace to JSON output.
Can be TRUE/FALSE or a number specifying the number of spaces to indent.}
}
\description{
currently just a wrapper around jsonlite::toJSON
}
