% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{write_json}
\alias{write_json}
\title{write a JSON file, a wrapper around jsonlite::write_json with different defaults}
\usage{
write_json(data, path, pretty = TRUE, auto_unbox = TRUE, ...)
}
\arguments{
\item{data}{the data to write}

\item{path}{the path of the JSON config file}

\item{pretty}{adds indentation whitespace to JSON output.
Can be TRUE/FALSE or a number specifying the number of spaces to indent.}

\item{auto_unbox}{whether if the vectors of 1 element shoudl be written as a scalars,
i.e \verb{"string" instead of ["string"]}}

\item{...}{passed to \code{\link[jsonlite:read_json]{jsonlite::write_json()}}}
}
\description{
write a JSON file, a wrapper around jsonlite::write_json with different defaults
}
\keyword{internal}
